* add initial Mac OS X Retina Display support 	* add CDrawContext::Transform to transform any of the CDrawContext draw commands 	* add CDrawContext::drawLine and mark the old methods (moveTo, lineTo) as deprecated 	* add CDrawContext::drawLines which takes a vector of point pairs and mark the old method as deprecated 	* add CDrawContext::drawPolygon which takes a vector of points and mark the old method as deprecated

+++ trunk/vstgui/lib/cviewcontainer.h
@@ -232,7 +232,7 @@ public:
 			}
 			it++;
 		}
-		return result.size ();
+		return static_cast<uint32_t> (result.size ());
 	}
 
 	//-------------------------------------------
